refactor: useSearchData 추가 state 통합
